14. 先頭からN行を出力
